fix: stx_mode 类型越界写 + 加 .gitignore 防测试产物入库
## 一、修掉一个真实的未定义行为（告警门禁抓到）

门禁在 `preload.c:2801` 报出 `-Wincompatible-pointer-types`：

```
warning: passing argument 3 of 'l2s_rt_patch_statx_full'
         from incompatible pointer type
```

**不是误报**，`struct statx` 的字段宽度不一样：

| 字段 | 类型 | 大小 |
|---|---|---|
| `stx_mask` | `__u32` | 4 |
| `stx_nlink` | `__u32` | 4 |
| **`stx_mode`** | **`__u16`** | **2** |

而函数签名把 `stx_mode` 声明成 `unsigned int *`，于是内部这一行

```c
*stx_mode = (*stx_mode & ~S_IFMT) | S_IFREG;
```

是一次 **4 字节写**，覆盖 `stx_mode`(offset 28-29) **+ `__spare0`(30-31)**。

**为什么不致命但必须修**：覆盖的恰好是 spare 字段，所以实测行为「看起来正常」。
但它依赖「后面正好是 spare」这个**巧合**，不是 ABI 契约 —— 一旦内核在 30-31
放了别的东西，就会静默写坏客户的结构体。而且这是编译器明确报告的未定义行为。

修法：把 `stx_mode` 参数改为 `uint16_t *`（跨 4 处：`l2s-runtime.h` 签名、
`l2s-runtime.c` 的 `patch_statx_impl` 与 `full`、`syscall_guard.c` 的前向声明
与调用点）。写入改为精确的 2 字节。

`syscall_guard.c` 的调用点原本用 `&base[7]`（u32 下标 7 = offset 28 的
**4 字节视图**），改为字节寻址 + `uint16_t *`：

```c
uint16_t *mode16 = (uint16_t *)(void *)((unsigned char *)(uintptr_t)a4 + 28);
```

字段偏移经 `offsetof` 实测确认（`stx_mask=0`、`stx_nlink=16`、`stx_mode=28`），
不是推断。

**修后两条路径仍一致**：
```
statx(291):  nlink=2 mode=0100600 S_ISLNK=0
stat:        nlink=2 mode=0100600 S_ISLNK=0
```

★ 这个缺陷**是告警门禁抓到的** —— 而门禁本身在早前也是一处被修复的漏洞
（它曾用 `-fsyntax-only` 漏报所有后端告警）。这次是它的直接回报。

## 二、加 `.gitignore`（把"手工核对"变成默认安全）

项目在开发中会**大量产生**测试产物，且与源码同目录：
```
src/runtime/t_null  src/runtime/t_write  src/runtime/t_stack
src/proc/test_proc  src/proc/test_proc.ub
```
实测一次 `git add -A` 前存在 5 个 ELF 可执行文件，当时靠**手工核对**才发现。
仓库此前**没有 `.gitignore`**，`git add -A` 很容易把它们一起提交 ——
几十个几 MB 的二进制进仓库，且没人会看 diff。

规则按名字逐个列出而非用宽通配符（`test_*` 会误伤 `test_*.c` 源文件）。

**验证（这一步不能省）**：
```
源文件总数: 40
✅ 没有任何源文件被忽略
✅ src/runtime/t_null        已忽略
✅ src/proc/test_proc        已忽略
✅ test/test_fakeroot        已忽略
✅ build/libbxroot-runtime.so 已忽略
```

## 三、目标维度的完整验收

**proroot 全功能**：
| 能力 | 结果 |
|---|---|
| `dsh --version` | `0.1.5-rc.2` |
| `dsh --help` | rc=0 |
| 子进程 `spawn` | `status=0 out="ok\n"` |
| 经 shell 的 spawn | `status=0 out="via-shell\n"` |
| `execSync` | `OK out="exec-ok"` |
| **`dsh web` 启动** | ✅ 打印 URL + token |
| **插件 `Cannot find package`** | **0 个** |
| HTTP 响应 | 303（鉴权重定向，正常） |
| **l2s 端到端契约** | **PASS**（`nlink=2 islink=false`，与官方逐项一致） |

**proot 兼容性**：
| 项 | 结果 |
|---|---|
| CLI 兼容（38 项，含 DSHA 真实命令行） | 38/38 PASS |
| `PROOT_TMP_DIR` | ✅ |
| `PROOT_VERBOSE`（数字语义，非布尔） | ✅ |
| `PROOT_IGNORE_MISSING_BINDINGS` | ✅ |
| CLI 覆盖率 | 已支持 27 / 明确拒绝 7 / **未处理 0** |
| DSHA 真实参数验证 | 被接受且执行成功（9 条 bind） |

## 回归

```
sh test/RUN_ALL.sh  →  12 通过 / 0 失败（无已知缺陷）
告警门禁            →  11 个编译单元零告警
```

SPDX-License-Identifier: MIT

+++ src/l2s/l2s-runtime.c
@@ -617,7 +617,7 @@ void l2s_rt_patch_stat(struct stat *st, const char *path)
 
 /* 公共实现：stx_mode 为 NULL 时只补 nlink（历史行为，向后兼容）。 */
 static void patch_statx_impl(unsigned int *stx_nlink, unsigned int *stx_mask,
-                             unsigned int *stx_mode,
+                             uint16_t *stx_mode,
                              unsigned int statx_nlink_bit, const char *path)
 {
     char mid[L2S_PATH_MAX];
@@ -659,8 +659,12 @@ static void patch_statx_impl(unsigned int *stx_nlink, unsigned int *stx_mask,
      * l2s_rt_patch_stat() 早就在抹 S_IFLNK 了；statx 只是同一个语义
      * 的现代接口，没有理由区别对待 —— 此前的差异纯属遗漏。
      */
+    /*
+     * 精确的 2 字节写（stx_mode 是 __u16）。见 l2s-runtime.h 的类型说明 ——
+     * 早前按 4 字节写会越界覆盖 __spare0，属未定义行为。
+     */
     if (stx_mode != NULL && g_hide_symlink)
-        *stx_mode = (*stx_mode & ~(unsigned int)S_IFMT) | (unsigned int)S_IFREG;
+        *stx_mode = (uint16_t)((*stx_mode & ~(uint16_t)S_IFMT) | (uint16_t)S_IFREG);
 
     g_stats.nlink_patched++;
 }
@@ -683,7 +687,7 @@ void l2s_rt_patch_statx(unsigned int *stx_nlink, unsigned int *stx_mask,
  * 原有的 4 参数版本保留，是为了让调用方可以分步迁移、不必一次改两处。
  */
 void l2s_rt_patch_statx_full(unsigned int *stx_nlink, unsigned int *stx_mask,
-                             unsigned int *stx_mode,
+                             uint16_t *stx_mode,
                              unsigned int statx_nlink_bit, const char *path)
 {
     patch_statx_impl(stx_nlink, stx_mask, stx_mode, statx_nlink_bit, path);
